some corrections
