+++ app/users/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 from app.users.models import User
 from django.contrib.auth.forms import UserCreationForm
-from app.core.constants import UserRoleE
+from app.users.constants import UserRoleE
 
 
 class UserRegisterForm(UserCreationForm):
@@ -9,4 +9,13 @@ class UserRegisterForm(UserCreationForm):
 
     class Meta:
         model = User
-        fields = ['username', 'email', 'role', 'phone_number', 'password1', 'password2']
+        fields = ['username', 'email', 'phone_number', 'password1', 'password2']
+
+
+class UserAdminCreateForm(UserCreationForm):
+    email = forms.EmailField
+
+    class Meta:
+        model = User
+        fields = ['username', 'email', 'phone_number', 'role', 'password1', 'password2']
+
