+++ app/users/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render, redirect
-from app.users.forms import UserRegisterForm
+from app.users.forms import UserRegisterForm, UserAdminCreateForm
 from django.contrib import messages
 
 
@@ -14,4 +14,18 @@ def register(request):
             return redirect('login')
     else:
         form = UserRegisterForm
+    return render(request, 'users/register.html', {'form': form})
+
+
+def create_admin(request):
+
+    if request.method == 'POST':
+        form = UserAdminCreateForm(request.POST)
+        if form.is_valid():
+            form.save()
+            username = form.cleaned_data.get('username')
+            messages.success(request, 'Admin Account has been created succesfully')
+            return redirect('login')
+    else:
+        form = UserAdminCreateForm
     return render(request, 'users/register.html', {'form': form})
