+++ templates/app/base.html
@@ -31,18 +31,18 @@
             {% if user.is_authenticated %}
             <a class="nav-item nav-link">Hello {{user.get_username}} !</a>
             {% if user.is_superuser %}
-            <a class="nav-item nav-link" href="{% url 'register' %}">+ACCOUNT</a>
+            <a class="nav-item nav-link" href="{% url 'create-admin' %}">+ACCOUNT</a>
             {% endif %}
             <a class="nav-item nav-link" href="{% url 'logout' %}">Logout</a>
             {% else %}
             <a class="nav-item nav-link" href="{% url 'login' %}">Login</a>
             <a class="nav-item nav-link" href="{% url 'register' %}">Register</a>
             {% endif %}
-            {% if user.role is 2 %}
+            {% if user.role == 2 %}
             <a class="nav-item nav-link" href="#">SELL</a>
             <a class="nav-item nav-link" href="#">PRINT TICKETS</a>
             {% endif %}
-            {% if user.role is 3 %}
+            {% if user.role == 3 %}
             <a class="nav-item nav-link" href="#">+HALL</a>
             <a class="nav-item nav-link" href="#">+MOVIE</a>
             <a class="nav-item nav-link" href="#">+SHOW</a>
