+++ templates/users/logout.html
@@ -8,7 +8,7 @@
 <h2> You have been logged out </h2>
 	<div class="border-top pt-3">
 		<small class="test-muted">
-			Need An Account?<a class=ml-2 href="{% url 'login' %}">Log In Again</a>
+			Need An Account?<a class=ml-2 href="{% url 'register' %}">Sign Up</a>
 		</small>
 	</div>
