+++ templates/users/password_reset.html
@@ -25,7 +25,7 @@
 	</form>
 	<div class="border-top pt-3">
 		<small class="test-muted">
-			Need An Account?<a class=ml-2 href="{% url 'login' %}">Sign In</a>
+			Need An Account?<a class=ml-2 href="{% url 'register' %}">Sign Up</a>
 		</small>
 	</div>
 </div>
