+++ templates/users/register.html
@@ -27,7 +27,7 @@
 	</form>
 	<div class="border-top pt-3">
 		<small class="test-muted">
-			Already Have An Account?<a class=ml-2 href="{% url 'login' %}">Log In</a>
+			Already Have An Account?<a class=ml-2 href="{% url 'login' %}">Sign In</a>
 		</small>
 	</div>
 </div>
